Initial boilerplate code
Implements a basic project structure for opendigitizer.
The client only shows the implot and imgui samples and for now only
compiles to emscripten.
The service part includes an endpoint for distributing the wasm client
on the http backend and provides workers for the acquisition and
flowgraph properties.
At the moment these are commented out because they immediately crash
the service.