Implement fetching dashboards from a remote source
opendigitizer now provides a '/dashboard/ service. When passing a url
to opendigitizer-ui, e.g. http://localhost:8080/dashboard, as a command
line argument it will load the dashboard from there. The same can be
done on emscripten by appending the dashboard url as a query argument
to the application url, like so:
http://localhost:8000/?http://localhost:8080/dashboard