Fix occasional failure to load remote dashboards
When loading remote dashboards there code was doing two requests, one
for the flowgraph and one for the dashboard. However, for correct loading
the flowgraph must be parsed before the dashboard, but nothing was making
sure that happened.
Now  instead request both the flowgraph and dashboard YAML trees with a
single request, so that we have a single callback in which we first parse
the flowgraph and then the dashboard.