Implement a side panel with block controls
Clicking on a block in the flowgraph ones a side panel with controls
to change the block settings. Likewise, clicking on a data signal in a plot
legend opens a similar panel where it's possible to navigate back and forth
in the blocks chain, add new blocks in the chain and change blocks settings.
The panels close after 15 seconds of inactivity.