CMake: remove check of defined target
Since the target was defined on subsequent cmake runs, the build
would only work for the first cmake run and not for subsequent runs.
The cmake structure will eventually need to be cleaned up properly,
this is just to get everything working again.
Originally the idea was that both subprojects could be built
individually from different ide projects to allow switching to wasm
easily, while the toplevel cmake should be used by non-dev uses like
CI and people who just want to build the state of the repository with
a single command.
This is not possible anymore easily with the additional subprojects
that both depend on.