Adapt to graph-prototype changes
As the sinks now register themselves only when started, we need an
initial delay in processing to get all data in the test listener.
Use gr::testing::Delay for that.

That also removes the need for the workaround for sinks being mixed
up in simultaneously existing gr::Graphs. That's fine now, as long
only one graph is executed by a scheduler, at a time.

To force-stop the scheduler, use message ports instead of polling
isProcessing.