Frank/metadata propagation (#164)
* RemoteSource: Forward remote signal metadata

* GnuRadioWorker: Fix metadata handling

The automatic forwarding uses the shortKey (e.g. "signal_name") instead
of they key ("key::signal_name").

* CountSource: Remove initial delay

This can now be achieved with gr::testing::Delay.

* Adapt to graph-prototype changes

As the sinks now register themselves only when started, we need an
initial delay in processing to get all data in the test listener.
Use gr::testing::Delay for that.

That also removes the need for the workaround for sinks being mixed
up in simultaneously existing gr::Graphs. That's fine now, as long
only one graph is executed by a scheduler, at a time.

To force-stop the scheduler, use message ports instead of polling
isProcessing.

* Do not spam console for unknown signals