Implement signal multi-selection
This way they can be added with fewer clicks.
The latest Imgui version is required for this, as multi-selection has
only recently been implemented in imgui:
https://github.com/ocornut/imgui/issues/6518

Fixes #193

While we are at it, this patch also implements clipping of the list with
an ImGuiListClipper.
This is possible, because all rows have the same height in the table.
With this change, scrolling through a large list (100k+ rows) is
significantly smoother.

Signed-off-by: Magnus Groß <magnus.gross@kdab.com>