CI: update github pages action versions
This commit updates the github pages actions to versions which use
supported nodejs versions and are generally up to date.
This fixes all the warnings in the github actions panel.

There is one change which broke the CI, as with the new versions
uploading the same artifact name twice will not overwrite but create an
error. This is fixed by only uploading the Release version.

Also allows the special branch `fixOnlineDemo` to be used to test github
pages deployment without having to merge the PR, by also publishing it
to github pages.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>