Integrate ImGui Test Engine into the build
ImGui itself needs to be built with special flags, for this reason
this is guarded behind a ENABLE_IMGUI_TEST_ENGINE option, defaults
to ON for developers and OFF for release. Can be overriden.
Set it true for CI and developer environments, false for
production users.

Pass -DENABLE_IMGUI_TEST_ENGINE=ON to CI

Since ImGui depends on the test engine and the test engine depends
on ImGui, the recommended way is to bake all the .o into the same
target, instead of having a dedicated Test Engine library.

fix cmake: Move enable_testing() before all add_test() calls
Otherwise ctest won't pick them.

For task #208

cmake: Calculate a good default for ENABLE_IMGUI_TEST_ENGINE

Signed-off-by: Sergio Martins <sergio.martins@kdab.com>