Integrate ImGui Test Engine into the build
ImGui itself needs to be built with special flags, for this reason
this is guarded behind a ENABLE_IMGUI_TEST_ENGINE option, default
false. Set it true for CI and developer environments, false for
production users.

Since ImGui depends on the test engine and the test engine depends
on ImGui, the recommended way is to bake all the .o into the same
target, instead of having a dedicated Test Engine library.

For task #208

Signed-off-by: Sergio Martins <sergio.martins@kdab.com>