Add proof of concept GUI unit-test
PopupMenu chosen at random just for the purpose of showing how
testing works, it's not a complete popup test.

Introduces ImGuiTestApp, a way to share SDL and ImGui init code
across all future tests.

For end-to-end tests we'll need to factor-out parts of the application
to make it reusable by tests. Out of scope for #208, as this is just
about cmake integration and minimal example.

To be decided:
 - Offscreen rendering (no window) would be a nice to have, so
developers don't get spammed with windows when running 'ctest -j12'

- Use boost::ut for printing test results.
Both ImGui Test Engine and boost::ut have facilities for asserts
and result reporting, we'll settle for boost::ut, which is already
being used for other tests and more evolved.

- Lot's of settings to tune later as we gain experience. For example,
mouse teleportation can speed up tests, but might not be great if
the component being tested uses timeouts internally (popup menus for
example).

- Coding style: ImGuiTestApp is not header-only and is polymorphic.
Since it's not part of the production build I don't think it needs
to be micro-optimized and also since we'll have dozens of test
exacutables eventually.

Finishes task #208

Signed-off-by: Sergio Martins <sergio.martins@kdab.com>