Reenable picoscope block in opendigitizer-service
The digitizer block was disabled due to gr-digitizers not following some
of the changes in gnuradio4. This was fixed in gr-digitizers#147 which
refactored the picoscope blocks to the current block API but was never
followed up in opendigitizer.

I had to change the data type for all blocks to float because the
picoscope block does not support double anymore. It should be possible
to follow up these changes and add a converter block to use double in
the rest of the flowgraph.

I also removed some emscripten guards since the opendigitizer-service
part is not supposed to build on emscripten anyway.

This changes depend on reading the template parameters for the blocks
from the yaml.

With these changes it is possible to run with the picoscope again, I
still cannot get data via the subscription, probably due to some channel
name missmatches which i still try to resolve.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>