Picoscope flowgraph: change types and channel settings
Followup for changes in gr-digitizer which limit the output types of the
digitizer to float, unint16 and UncertainValue<float>.
Due to the http subscriptions only working for double values (to be
investigated why) the values are converted to double using a converter
block.

Also changes the name of channelA to test, s.t. the same dashboards as
for the builtin test flowgraph can be used.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>