UI: Add picoscope4000 block to native ui build
This allows to test with a locally connected picoscope.

Compiles but is currently not working because the UI does not support
the use of sub indices to connect ports in port collections.
Probably not worth fixing this in opendigitizer-ui as effort should be
concentrated on using upstream gr4 facilities here.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>