Frank/dataset worker (#226)
* GnuRadioWorker: Allow subscription to DataSetSink

Allow clients to subscribe to DataSetSink data, using the
acquisitionModeFilter=DataSet query parameter. Each DataSet is
converted to an Acquisition object. As acquisition only supports
a single signal at a time, only a single signal name can be passed
via acquisitionModeFilter, which is then extracted from the dataset.

* GnuRadioWorker: Use advanced tag matching

For the Triggered, Multiplexed and Snapshot acquisition modes, use
trigger::BasicTriggerNameCtxMatcher for more advanced matching of
trigger tags (TimingCtx, etc.).

* Add RemoteDataSetSource

Add source that treats the incoming data as a data set and outputs
it as such. Rename the existing RemoteSource to RemoteStreamSource.

* Add new example dashboard demoing the remote data set source

---------

Signed-off-by: Frank Osterfeld <frank.osterfeld@kdab.com>