App: manage scheduler state from the main thread
This is required for the emscripten fetch calls to be issued in the main
browser thread. If they are issued on another thread, they are only
processed once the thread is finished.

Otherwise we would have to investigate if there is a way to control
on which thread the callbacks should be processed or we would have to
proxy the emscripten_fetch calls to the main thread.
This is further complicated because it needs to be transparent to the
part of the code that also compiles to native.

Set graph threadpool size to one

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>