Investigating emscripten_fetch for remote_source
as the current behavior is very broken, (relies on state changes being
called from the main application thread, otherwise fetch callbacks will
not be invoked) this tries to move all the fetch logic to the scheduler
thread pool.

The initial idea was to manually invoke `emscripten_sleep(0)` to yield
control to tbe browser's event processing, but this doesn't ssem to work
and the application crashes as soon as emscripten_sleep is invoked.
Hard to see what is actually happening, but my theory is that it also
invokes some SDL/GL functionality, which is only allowed to be called in
betweeen render passes.

A possible alternative could be to run the fetch call synchronously on a
separate worker thread. The disadvantage is,tthat this would need a
sparate exclusive thread (which are limited to a fixed number at compile
time for emscripten) for each running subscription.