Remove usage of PushOverrideID when changing between view/layout
Because we don't need it anymore and because we can't.

Dockspace need to have a stable ID, if its parent ID changes then
dock widgets appear all out of place.

We don't need it because the plots are now inside their own window
(ImGui::Begin), they have a stable ID, and zoom is preserved

Signed-off-by: Sergio Martins <sergio.martins@kdab.com>