FileServerRestBackend: set caching header
Send the cache-control header field to allow browsers to cache the
static file artifacts.
This seems to work for most resources, the wasm file (167MB) exceeds the cached
filesize (~50MB in my firefox configuration) and is not cached, but all
the index.worker.js requests now directly go to cache.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>