FileServerRestBackend: use read instead of istreambuf
Serving the 160mb wasm file took ~3min on from the deployed system.
Adding some logging output showed, that most of that time was spent
between having obtained the size of the file and copying it into a
buffer to pass to httplib.
Might be caused by different backing filesystem or just different
performance of the development vs deployment systems.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>