RemoteSource: add default port and only subscribe once host is set.
Only subscribe to the host once the correct host is set.
Also simplifies the logic for stopping subscriptions by actively
tracking the active subcscription topic.
Adds a port to the default hostname to be assumed if the scheduler has
not (yet) set the base url. This also allows the native client to
subscribe to remote sources.

Had to increase the stack size for the proxying to the main thread to
work.

Also adds a 5s delay before trying to reconnect.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>