DasbhoardWorker: load default dashboards from directory
The directory can be configured via the DIGITIZER_REMOTE_DASHBOARDS
environement variable, which defaults to the relative path from a build
in the subdirectory of the project. This should make this change
transparent for most command line usages, but deployment will have to be
adapted to set the new location and copy the dashboard files.

Before they were compiled into the binary with cmakeRC, which required
frequent recompiles and made managing application specific flowgraphs
and dashboars alongside each other outside of this repo impossible.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>