Add block navigation in the blocks property panel
This makes it easy to see only the immediate neighbours.

We have very little space in the properties panel, thus the design
needs to be a bit different than the main FG:
  - Don't show port name (use tooltip instead)
  - Blocks are smaller
  - Less text shown (just the title)
  - Possibly we'll want word wrap
  - All goes inside an horizontal scroll area

The ports are colored, as suggested. But not sure which types should
use which colors. Red ports for now.

The styling is separated from the business logic, which is hidden
by a callback and context object.

For issue #202

Signed-off-by: Sergio Martins <sergio.martins@kdab.com>