Fix data race in qa_chart_fgdipole
The test was hanging randomly depending on luck.

The OD Scheduler already creates a thread which immediately
changes state to INITIALIZED and then to RUNNING. However, our
test was also starting the scheduler manually, which could race
and move the scheduler to RUNNING while its init() method hadn't
completely finished.

* fixed fg_dipole_intensity_ramp.grc to include explicit start/stop filter criteria for the StreamToDataSet blocks.
  - TODO: find a way to detect/filter when a new BPC starts (nominally via BPCTS and/or "CMD_BP_STOP" event

Signed-off-by: Sergio Martins <sergio.martins@kdab.com>
Signed-off-by: rstein <r.steinhagen@gsi.de>