Don't leave "selected block" pointer dangling when it's removed
When a block is selected we open the right properties sidepane.
When a block is removed by the user we should close the sidepane
otherwise it crashes.

The easiest way to do this is letting the graph model store it,
so it can also clear it without adding more callbacks.

Signed-off-by: Sergio Martins <sergio.martins@kdab.com>