ui: GraphModel: comment out allocation heavy code
This method was responsible for more than 2/3 of the processing time
during unresponsive load spikes as observed in firefox profiler.
Most of this was ultiately allocation of temporary heap objects, as the
lookup in the map has to construct the temporary key strings with the
colon and also the returned values.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>