Service: add second RestBackend for plain http
As openssl seems to add considerable performance overhead for data-heavy
subscriptions, add a second endpoint to provide plain http
subscriptions.
For now this will only work with the native client, to work with the
emscripten client from a browser, the subscription endponints will have
to set the correct cross origin headers (opencmw change).

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>