RemoteSource: Do not publish inacurate timestamps
Timestamps for dropped updates would be based on the inacurate
acquisitionTime field of the acq object (currently set as the first
tag's time since the sample rate needed to determine the correct time is
not accessible inside of the worker).

This filters out all timestamps equal to zero to allow publishing tags
without any time attached.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>