qa_GnuRadioWorker: Fix test with coverage
With enabled coverage, the timing of the flowgraph execution changed
slightly which broke the unittest.
The main problem is that updating the metadata for the dns update hook
was implemented via the message system. Due to the relatively short
subscriptions which were often processed in a single processWork
invokation, the settings-changed message that was emmited due to processing
propagated tags updating the block settings did not reach the scheduler
before it was torn down again, leading to the dns metadata reporting the
default units and quantities of the sink block instead of the ones
configured in the source blocks.

This commit fixes these issues in different ways:
- increase the number of processed samples to something bigger than the
  buffer size to make sure multiple process work iterations are run
- add some small timeouts after the required samples have been received
  to allow the message processing to take place
- add a get request to all sink block settings after the setting
  subscriptions to ensure that we cannot miss updates happening between
  starting the flowgraph and the subscription message being processed.

This seems to be enough to fix the immediate issue, but this should be
refactored properly once the new message based API is complete.

Some additional changes that are mostly unrelated:
- load custom thread pools
- remove commented out code and outdated comments
- add a missing check for quantities in the metadata

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>