Scheduler: use messages instead of state access
Changes state changes of the scheduler to be triggered via messages
instead of directly updating the state. This is dangerous because
changing the state triggers emitting an update message from this thread
as well as from the scheduler thread in it's process Messages call.
Since these both use the same buffer writer from different threads this
triggers an abort if one thread tries to reserve space in the buffer
while another has not yet published its WriterSpan.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>