Add external input and output blocks to UiGraphModel
The code provides new accessors for iterating over the
graph's model, so that we can easily "inject" custom
blocks into the node editor rendering.

It's also important to these blocks to have some
of their properties filled as the code expect
types, direction, names, and lastly the backend
will need to approve the links between blocks,
and store/provide them when starting.