GHA: upload github pages even if some jobs fail
We want to upload to github pages even if for example the debug build
fails.
Unfortunately it is not possible to depend on only on the build result
of a specific entry in the build matrix:
https://github.com/orgs/community/discussions/42335

To emmulate this, the upload job is run unconditionally and the build
action sets an output variable if it successfully uploaded the artifact.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>